pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "readme-score"
version = "0.1.0"
description = "README.mdの品質をスコアリングするCLIツール"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "markdown-it-py>=3.0",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.scripts]
readme-score = "readme_score.cli:main"

[tool.setuptools.packages.find]
include = ["readme_score*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
